requirements.txt
flask
flask-cors
pysmi
pyyaml
boto3
newrelic